feat(os/bcm2837): GPIO driver, and route the UART pin mux through it
54 pins, eight functions each, selected three bits at a time across six
GPFSEL registers. Set and clear are separate write-only registers rather
than a read-modify-write on a level register, which is what lets a
single pin be driven without a lock -- worth preserving once there is
more than one core running.

The pull-up/pull-down control is a clocked sequence, not a register
write: drive GPPUD, wait, clock it into the target pins with GPPUDCLK,
wait, release. The waits are required by the peripheral spec and
skipping them silently leaves pins in their previous state.

uart.c now calls this instead of poking GPFSEL/GPPUD itself. That keeps
one owner of the pin mux, removes the duplicated clocked-pull sequence,
and means every boot exercises the GPIO driver before anything else
depends on it.

The self-test only READS function selects, and only for pins this kernel
already configured. Driving an arbitrary pin as a test would be reckless
on real hardware: header pins are wired to whatever the owner plugged
in, and asserting an output could short a driven line. Reading the mux
back proves it took, which is the part that is otherwise invisible.

Also routed the MMIO accessor macros in gpio.c, uart.c and mailbox.c
through uintptr instead of letting a peripheral address be an int on its
way to a pointer. That is precisely the 64-bit hazard class the os/port
import is going to be full of, so it should not be seeded here.

Intended end state is a 9P server presenting pins as files, so GPIO is
scriptable from sh and reachable by agents without an API -- that needs
the Dis VM, so the driver stays deliberately thin until then.

Suite 19/19.

Refs: INFR-404

+++ os/bcm2837/mailbox.c
@@ -25,7 +25,7 @@
 #include "io.h"
 #include "fns.h"
 
-#define MBOX(r)	(*(volatile u32int*)(MBOXREGS + (r)))
+#define MBOX(r)	(*(volatile u32int*)((uintptr)MBOXREGS + (r)))
 
 /*
  * Shared request buffer.  Single-threaded for now; when there is a
